feat/SD-51-pay-order-with-sea-labs-pay

See merge request reivaldo.julianto/seadeals-backend!75

+++ dto/get_product_variant_res.go
@@ -51,6 +51,6 @@ func (_ *GetProductVariantRes) From(pv *model.ProductVariantDetail) *GetProductV
 		Variant2Value: variant2Value,
 		VariantCode:   pv.VariantCode,
 		PictureURL:    pictureURL,
-		Stock:         pv.Stock,
+		Stock:         uint(pv.Stock),
 	}
 }
